% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_crossover.R
\name{ccDesign}
\alias{ccDesign}
\title{Create a list of stratification parameters}
\usage{
ccDesign(...)
}
\arguments{
\item{...}{Additional arguments to customize the stratification parameters. These arguments are used to update the
default parameters.}
}
\value{
A list containing the stratification design parameters. The list includes:
\itemize{
\item strat_vars: character or NULL. Variables to stratify the data.
\item time_var: character or NULL. Time variable.
\item time_lag: integer. Lag used to stratify data (default is 7).
\item time_size: integer. Size of strata (default is 4).
\item scheme: character. Stratification scheme (default, and sole implemented scheme is "time stratified").
}
}
\description{
This function creates a list of design parameters used in stratification. The default parameters can be
customized by passing arguments to the function.
}
