% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{getNewXA}
\alias{getNewXA}
\title{Post-fit design matrices constructor.}
\usage{
getNewXA(terms, df)
}
\arguments{
\item{terms}{The output `terms` of `hlnm`.}

\item{df}{A data.frame with all the necessary columns (as in the `data` input of `hlnm`).}
}
\value{
A list containing X and A, the fixed and random effects design matrices.
}
\description{
Replicates the loop in `hlnm` that construct the design matrices, without modifying `terms` 
and creating the other objects (i.e., the precision matrices).
}
