% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hnlm_helpers.R
\name{hnlm_helpers}
\alias{hnlm_helpers}
\alias{collectTerms}
\alias{getExtra}
\alias{getDesign}
\alias{getGammaSetup}
\alias{getPrecision}
\alias{getThetaSetup}
\alias{addFPoly}
\alias{addRPoly}
\title{Helpers for the \code{hnlm} function}
\usage{
collectTerms(formula)

getExtra(term, data, cc_matrix)

getDesign(term, data)

getGammaSetup(term)

getPrecision(term)

getThetaSetup(theta_info, term)

addFPoly(term)

addRPoly(term)
}
\description{
This page provides common information for functions that handle the setup and
processing of terms used in the model.
These functions include:
\itemize{
\item \code{collectTerms()}: Sets up the terms object, which is the core of the model.
\item \code{getExtra()}: Constructs additional quantities specific to an effect.
\item \code{getDesign()}: Constructs the design matrix for a given effect.
\item \code{getGammaSetup()}: Sets up the gamma parameters specific to an effect.
\item \code{getPrecision()}: Constructs the precision matrix for a random effect.
\item \code{getThetaSetup()}: Sets up theta information for a specific effect.
\item \code{addFPoly()}: Adds fixed polynomial effects to the model.
\item \code{addRPoly()}: Adds random polynomial effects to the model.
}
}
