% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hnlm_helpers.R
\name{hnlm_helpers}
\alias{hnlm_helpers}
\alias{collectTerms}
\alias{getExtra}
\alias{getDesign}
\alias{getGammaSetup}
\alias{getPrecision}
\alias{getThetaSetup}
\alias{addFPoly}
\alias{addRPoly}
\title{Helpers for the `hnlm` function}
\usage{
collectTerms(formula)

getExtra(term, data, cc_matrix)

getDesign(term, data)

getGammaSetup(term)

getPrecision(term)

getThetaSetup(theta_info, term)

addFPoly(term)

addRPoly(term)
}
\description{
This page provides common information for functions that handle the setup and 
processing of terms used in the model.
These functions include:
- `collectTerms()`: Sets up the terms object, which is the core of the model.
- `getExtra()`: Constructs additional quantities specific to an effect.
- `getDesign()`: Constructs the design matrix for a given effect.
- `getGammaSetup()`: Sets up the gamma parameters specific to an effect.
- `getPrecision()`: Constructs the precision matrix for a random effect.
- `getThetaSetup()`: Sets up theta information for a specific effect.
- `addFPoly()`: Adds fixed polynomial effects to the model.
- `addRPoly()`: Adds random polynomial effects to the model.
}
